+++ docs/release-notes.md
@@ -37,6 +37,8 @@ and
 - Added `Arbitrary.tuple1(), Arbitrary.tuple2(), Arbitrary.tuple3(), Arbitrary.tuple4()`
   to [generate tuples of same base type](/docs/snapshot/user-guide.html#tuples-of-same-base-types)
 
+- Added `Statistics.percentage(..)` and `StatisticsCollector.percentage(..)`
+
 #### Breaking Changes
 
 #### Bug Fixes